basic wall collision detection

+++ src/commonMain/kotlin/at/gnu/gridz/GridzInput.kt
@@ -2,5 +2,5 @@ package at.gnu.gridz
 
 interface GridzInput {
 
-    fun tick(inputX: Double, inputY: Double, dt: Long)
+    fun tick(inputX: Double, inputY: Double, dt: Float)
 }
